requirements.txt
python-nmap>=0.7.1
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0.0
jinja2>=3.0.0
reportlab>=4.0.0
requests>=2.31.0
pydantic>=2.0.0
langchain-ollama>=0.2.0
langchain>=0.3.0
